Very close to final push
Need help with two canvas removing one another
