+++ src/edu/macalester/comp127/FarmArt/Brush.java
@@ -1,7 +1,9 @@
 package edu.macalester.comp127.FarmArt;
 
+/**
+ * A interface brush
+ */
 public interface Brush {
 
-    void apply(Tile tile);
-    
+    void apply(Tile tile);  
 }
