+++ src/edu/macalester/comp127/FarmArt/Menu.java
@@ -3,14 +3,23 @@ package edu.macalester.comp127.FarmArt;
 import edu.macalester.graphics.CanvasWindow;
 import edu.macalester.graphics.Rectangle;
 import edu.macalester.graphics.ui.Button;
-import java.awt.Color;
  
+/**
+ * Represent a menu on the canvas with zoom in/out button, drag button, element button, and background button
+ * To allow users modifier their farm
+ */
 public class Menu {
 
-   private Boolean dragging = false;
    private Brush currentBrush;
    private Rectangle buttonIndicator;
+   private Boolean dragging = false;
  
+   /**
+    * Create a menu with its mouse drag and mouse drag 
+    * @param canvas
+    * @param tileManager
+    * @param buttonIndicator
+    */
    public Menu(CanvasWindow canvas, TileManager tileManager, Rectangle buttonIndicator) {
         this.buttonIndicator = buttonIndicator;
         addZoomButtons(canvas, tileManager);
@@ -35,6 +44,12 @@ public class Menu {
         });
    }
  
+   /**
+    * Add background button to the canvas with onClick animation
+    * @param canvas
+    * @param background
+    * @param y
+    */
    public void addBackgroundButton(CanvasWindow canvas, BackgroundType background, double y) {
        Button backgroundButton = new Button(background.getName());
        backgroundButton.setPosition(120, y);
@@ -46,11 +61,12 @@ public class Menu {
        });
    }
  
-   private void showSelectedButton(Button button) {
-       buttonIndicator.setPosition(button.getPosition());
-       buttonIndicator.setSize(button.getSize());
-   }
- 
+   /**
+    * Add element button to the canvas with onClick animation
+    * @param canvas
+    * @param elementType
+    * @param y
+    */
    public void addElementButton(CanvasWindow canvas, ElementType elementType, double y) {
        Button elementButton = new Button(elementType.getName());
        elementButton.setPosition(120, y);
@@ -62,6 +78,10 @@ public class Menu {
        });
    }
  
+   /**
+    * Add drag button to the canvas with onClick animation
+    * @param canvas
+    */
    private void addDragButton(CanvasWindow canvas) {
        Button drag = new Button("drag");
        drag.setPosition(70, 60);
@@ -72,6 +92,11 @@ public class Menu {
         );
     }
  
+    /**
+     * Add zoom in, zoom out, and standard button to the canvas with onClick animation
+     * @param canvas
+     * @param tileManager
+     */
    private void addZoomButtons(CanvasWindow canvas, TileManager tileManager) {
        Button zoomIn = new Button("+");
        Button zoomOut = new Button("-");
@@ -86,4 +111,13 @@ public class Menu {
        zoomOut.onClick(() -> tileManager.scaleDown());
        standard.onClick(() -> tileManager.scaleNormal());
    }
+
+    /**
+    * Set the buttonIndicator to the current selected button
+    * @param button 
+    */
+    private void showSelectedButton(Button button) {
+        buttonIndicator.setPosition(button.getPosition());
+        buttonIndicator.setSize(button.getSize());
+    }
 }
