+++ src/edu/macalester/comp127/FarmArt/Tile.java
@@ -5,13 +5,24 @@ import edu.macalester.graphics.Image;
 import edu.macalester.graphics.Rectangle;
 import java.awt.Color;
 
+/** 
+ * Represent a tile as graphicsGroup 
+ * A tile has a background color
+ * A tile has a element
+ */
 public class Tile extends GraphicsGroup{
     
     private final Rectangle background;
-    private ElementType currentElement;
+    private ElementType currentElementType;
     private Image elementImage;
     private String type;
     
+    /**
+     * A constructor of Tile
+     * @param x
+     * @param y
+     * @param size 
+     */
     public Tile(double x, double y, double size) {
         this.setPosition(x, y);
         background = new Rectangle(0, 0, size, size);
@@ -19,36 +30,44 @@ public class Tile extends GraphicsGroup{
         setBackgroundColor(null, null);
     }
 
+    /**
+     * Set the background color of a tile
+     * Controls tile's stroke and filled color
+     * @param color
+     * @param type
+     */
     public void setBackgroundColor(Color color, BackgroundType type) {
         if (color == null) {
             background.setFillColor(Color.white);
             background.setStrokeColor(Color.LIGHT_GRAY);
-        } else {
-            if (currentElement == null || currentElement.getName().contains(type.getName())) {
-                background.setFillColor(color);
-                background.setStroked(false);
-                this.type = type.getName();
-            }
+        } 
+        else {
+            background.setFillColor(color);
+            background.setStroked(false);
+            this.type = type.getName();
         }
     }
 
-    public void setElement(ElementType currentElement) {
-        Image newElement = new Image(currentElement.getName()+ ".png");
+    /**
+     * Set the element in a tile and updates the current element 
+     * @param currentElementType
+     */
+    public void setElement(ElementType currentElementType) {
+        Image newElement = new Image(currentElementType.getName()+ ".png");
         newElement.setMaxHeight(50);
         newElement.setMaxWidth(50);
-
-        if (currentElement.getType().contains(type) || currentElement.getType() == "all") {
+        
+        if (currentElementType.getType().contains(type) || currentElementType.getType() == "all") {
+            if (currentElementType.getName() == "blank") {
+                setBackgroundColor(null, null);
+            }
             if (elementImage != null) {
                 remove(elementImage);
             }
             newElement.setCenter(background.getCenter());
             add(newElement);
             this.elementImage = newElement;
-            if (currentElement.getName() == "blank") {
-                this.currentElement = null;
-            } else {
-                this.currentElement = currentElement;
-            }
+            this.currentElementType = currentElementType; 
         }
     }
 }
